+++ src/arch/armv8/inc/arch/iommu.h
@@ -20,7 +20,7 @@
 
 typedef struct {
     uint32_t global_mask;
-    int32_t ctx_id;
+    size_t ctx_id;
 } iommu_vm_arch_t;
 
 #endif
