+++ src/arch/armv8/vmm.c
@@ -30,10 +30,10 @@ void vmm_arch_init()
      * machine.
      */
 
-    static uint64_t min_parange = 0b111;
+    static size_t min_parange = 0b111;
     static spinlock_t lock = SPINLOCK_INITVAL;
 
-    uint64_t temp_parange = MRS(ID_AA64MMFR0_EL1) & ID_AA64MMFR0_PAR_MSK;
+    size_t temp_parange = MRS(ID_AA64MMFR0_EL1) & ID_AA64MMFR0_PAR_MSK;
     spin_lock(&lock);
     if(temp_parange < min_parange) {
         min_parange = temp_parange;
@@ -45,7 +45,7 @@ void vmm_arch_init()
     if (cpu.id == CPU_MASTER) {
         parange = min_parange;
         if (parange_table[parange] < 44) {
-            for (int i = 0; i < PT_LVLS - 1; i++) {
+            for (size_t i = 0; i < PT_LVLS - 1; i++) {
                 vm_pt_dscr->lvl_wdt[i] = vm_pt_dscr->lvl_wdt[i + 1];
                 vm_pt_dscr->lvl_off[i] = vm_pt_dscr->lvl_off[i + 1];
                 vm_pt_dscr->lvl_term[i] = vm_pt_dscr->lvl_term[i + 1];
